Fixed up Perceptron and PerceptronLearner classes
The most important change here was the stopping criteria. Because we
were only checking to see if the newAccuracy was less than or equal to
the old accuracy, it wasn't stopping soon enough (it would go for
hundreds of thousands of epochs regularly on the voting task). The fix
was simple (once I figured it out): find the difference/change in the
accuracy since last epoch and see if it's less than 0.1 (any small
threshold value would probably work). This way, if it's "improving
accuracy" but only slightly, it will still stop before it starts
overfitting the data or getting wildly off course.
Other than that, I prettied up the plotting a bit, added a second plot
for the misclassification rate as it changes through the epochs, added
labels, etc.
I also changed the weight vector in the Perceptron class to be a C#
style Property with a private setter and public getter. That way I could
move the printing of the final weights and things down into the
PerceptronLearner class.