This is the initial commit with the initial tools for this project.
All the python files are designed for python 2.7.

The contents of commit are as follows:

rules/* - The official .pdfs of the rules for each of the Dominion
expansions. They are copyright Rio Grande Games, and are unmodified from
their official download site.

dominion-card-list.txt - The list (semi-organized by set) of all Dominion
cards.

dominion.py - The python dominion package (we need to write and expand
this). This contains the class that represents a the state of a game of
dominion, at least from the context of the log parser. The log parser will
use this class to track the current game state, and extract our machine
learning features from it.

isotropic_parser.py - This is main parsing class. It is a work in
progress, but already extracts the game header from a log file.

log_downloader.py - This utility downloads and extracts dominion logs from
isotropic.org. By default, it will download everything from February 1,
2013 to the current date, and then extract them. It will cache log files,
so it can be run multiple times without duplicating work (however, if
files aren't completely downloaded and it is interrupted, it won't
know to download it again. Extractions work safely being interrupted in
the middle though, they just duplicate their work.)

feature-planning.txt - This has my initial thoughts on what features we
might want to use for machine learning.

test.html - The current test document for testing/developing the
isotropic parser.