Refactoring to allow for easier implementation of Backprop
I pulled LinearUnit into a separate file and created a ThresholdedUnit
that can take any kind of thresholding function (wrapped in a class
implementing the IThresholder interface).
I also made the PerceptronThresholder take a high and a low value for
what values it returns, with the defaults being 1 and 0, respectively.
This allows us to use 1 and -1 like in the book if we want to.