Implemented a SigmoidUnit and working on a Backpropagation Learner
I've realized that I'm going to need to have more classes in between,
like for each Layer, because I'm running into problems trying to handle
all of the different SigmoidUnits in the learner class. Basically, I
need to be able to cache the inputs and outputs of each node through the
network on each prediction run in order to train correctly.